Backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
joblib==1.3.2
scikit-learn==1.4.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.7
requests==2.31.0
tldextract==5.1.1